django-repo/NoiroseServer/requirements.txt
Django==4.1.7
pip==23.1.2
django-allauth==0.52.0
django-appconf==1.0.5
django-rest-auth==0.9.5
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
mysqlclient==2.1.1
numpy==1.24.3
oauthlib==3.2.2
scipy
requests==2.28.2
requests-oauthlib==1.3.1
django-cors-headers